feat: extract otlp-logger-derived code into dedicated shim file
Moves the adapted code from otlp-logger (createExporter,
createLogProcessor, getOtlpLogger, loadContext) into
lib/otlp-logger-shim.js with full license attribution.
The transport module imports getOtlpLogger from the shim.  The
otlp-logger package is removed from dependencies and replaced with
direct OpenTelemetry SDK packages.
